update readme and scripts
